api/requirements.txt
flask
flask_mysqldb
flask_cors
pydantic
PyMySQL
flask_sqlalchemy 
flask-migrate
python-dotenv
cryptography>=41.0.0